Feature: Readme complete
1. Added readme categories:
An explanation of the what the app does and how it works.
A link to the other repo
A link to both deployed sites
List of technologies used
List unsolved problems which would be fixed in future iterations.
Document your planning, process and problem-solving strategy
Link to wireframes and user stories
An embedded screenshot of the app
Set up and installation instructions for front end application